feature: Split AFJ demo for two cases: Main and DidCommV2
