+++ demo/src/BaseAlice.ts
@@ -1,0 +1,51 @@
+import type { ConnectionRecord } from '@aries-framework/core'
+
+import { BaseAgent } from './BaseAgent'
+import { greenText, Output, redText } from './OutputClass'
+
+export class BaseAlice extends BaseAgent {
+  public connected: boolean
+  public connectionRecordFaberId?: string
+
+  public constructor(port: number, name: string) {
+    super({ port, name, useLegacyIndySdk: true })
+    this.connected = false
+  }
+
+  protected async getConnectionRecord() {
+    if (!this.connectionRecordFaberId) {
+      throw Error(redText(Output.MissingConnectionRecord))
+    }
+    return await this.agent.connections.getById(this.connectionRecordFaberId)
+  }
+
+  protected async receiveConnectionRequest(invitationUrl: string) {
+    const { connectionRecord } = await this.agent.oob.receiveInvitationFromUrl(invitationUrl)
+    if (!connectionRecord) {
+      throw new Error(redText(Output.NoConnectionRecordFromOutOfBand))
+    }
+    return connectionRecord
+  }
+
+  protected async waitForConnection(connectionRecord: ConnectionRecord) {
+    connectionRecord = await this.agent.connections.returnWhenIsConnected(connectionRecord.id, { timeoutMs: 200000 })
+    this.connected = true
+    console.log(greenText(Output.ConnectionEstablished))
+    return connectionRecord.id
+  }
+
+  public async acceptConnection(invitation_url: string) {
+    const connectionRecord = await this.receiveConnectionRequest(invitation_url)
+    this.connectionRecordFaberId = await this.waitForConnection(connectionRecord)
+  }
+
+  public async exit() {
+    console.log(Output.Exit)
+    await this.agent.shutdown()
+    process.exit(0)
+  }
+
+  public async restart() {
+    await this.agent.shutdown()
+  }
+}
